creating the pages, separating into components Fixes #3
