+++ src/pages/create-trip/index.tsx
@@ -1,0 +1,172 @@
+import { ArrowRight, Calendar, MapPin, Settings2, UserRoundPlus } from "lucide-react";
+import { FormEvent, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { InviteGuestsModal } from "./invite-guest-modal";
+import { ConfirmTripModal } from "./confirm-trip-modal";
+
+
+export function CreateTripPage() {
+
+    const navigate = useNavigate();
+
+    const [isGuestsInputOpen, setIsGuestsInputOpen] = useState(false)
+    const [isGuestsModalOpen, setIsGuestsModalOpen] = useState(false)
+    const [isConfirmTripModalOpen, setIsConfirmTripModalOpen] = useState(false)
+
+
+    const [emailsToInvite, setEmailsToInvite] = useState([
+        'andreadventure@hotmail.com',
+        'aurora@hotmail.com'
+    ])
+
+    function openGuestInput() {
+        setIsGuestsInputOpen(true)
+    }
+
+    function closeGuestsInput() {
+        setIsGuestsInputOpen(false)
+    }
+
+    function openGuestModal() {
+        setIsGuestsModalOpen(true)
+    }
+
+    function closeGuestModal() {
+        setIsGuestsModalOpen(false)
+    }
+
+    function openConfirmTripModal() {
+        setIsConfirmTripModalOpen(true)
+    }
+
+    function closeConfirmTripModal() {
+        setIsConfirmTripModalOpen(false)
+    }
+
+    function addNewEmailToInvite(event: FormEvent<HTMLFormElement>) {
+        event.preventDefault()
+
+        const data = new FormData(event.currentTarget)
+        const email = data.get('email')?.toString()
+
+        if (!email) {
+            return
+        }
+
+        if (emailsToInvite.includes(email)) {
+            return
+        }
+
+        setEmailsToInvite([...emailsToInvite, email])
+
+        event.currentTarget.reset()
+    }
+
+    function removeEmailFromInvite(emailToRemove: string) {
+        const newEmaiList = emailsToInvite.filter(email => email !== emailToRemove)
+        setEmailsToInvite(newEmaiList)
+    }
+
+    function createTrip(event : FormEvent<HTMLFormElement>) {
+        event.preventDefault()
+
+        navigate('/trips/1234')
+    }
+
+    return (
+        <div className="h-screen flex items-center justify-center bg-pattern bg-no-repeat bg-center">
+            <div className="max-w-3xl w-full px-6 text-center space-y-10">
+
+                <div className="flex flex-col items-center gap-3">
+                    <img src="/logo.svg" alt="plann.er" />
+                    <p className="text-zinc-300 text-lg ">Convide seus amigos e planeje sua próxima viagem!</p>
+                </div>
+
+                <div className="space-y-4">
+                    <div className="h-16 bg-zinc-900 px-4 rounded-xl flex items-center shadow-shape gap-3 " >
+                        <div className="flex item-center gap-2 flex-1">
+                            <MapPin className="size-5 text-zinc-400" />
+                            <input disabled={isGuestsInputOpen} type="text " placeholder="Para onde voçê vai?" className="bg-transparent text-lg placeholder-zinc-400 outline-none flex-1" />
+                        </div>
+
+                        <div className="flex item-center gap-2">
+                            <Calendar className="size-5 text-zinc-400" />
+                            <input disabled={isGuestsInputOpen} type="text " placeholder="Quando?" className="bg-transparent text-lg placeholder-zinc-400 w-40 outline-none" />
+                        </div>
+
+                        <div className="w-px h-6 bg-zinc-800" />
+
+
+                        {
+                            isGuestsInputOpen ? (
+
+                                <button onClick={closeGuestsInput} className="bg-zinc-800 text-zinc-200 rounded-lg px-5 py-2 font-medium flex items-center gap-2 hover:bg-zinc-700" >
+                                    Alterar local/data
+                                    <Settings2 className="size-5" />
+                                </button>
+                            ) : (
+
+
+                                <button onClick={openGuestInput} className="bg-lime-300 text-lime-950 rounded-lg px-5 py-2 font-medium flex items-center gap-2 hover:bg-lime-400" >
+                                    Continuar <ArrowRight className="size-5" />
+                                </button>
+                            )
+                        }
+                    </div>
+
+                    {isGuestsInputOpen && (
+
+
+                        <div className="h-16 bg-zinc-900 px-4 rounded-xl flex items-center shadow-shape gap-3 " >
+                            <button onClick={openGuestModal} className="flex item-center gap-2 flex-1 text-left">
+                                <UserRoundPlus className="size-5 text-zinc-400" />
+                                {emailsToInvite.length > 0 ? (
+                                    <span className="text-zinc-100 text-lg flex-1">
+                                        {emailsToInvite.length} pessoa(s) convidada(s)
+                                    </span>
+
+                                ) : (
+                                    <span className="text-zinc-400 text-lg flex-1">Quem estará na viagem?</span>
+
+                                )}
+
+                            </button>
+
+
+
+                            <div className="w-px h-6 bg-zinc-800" />
+
+
+
+                            <button onClick={openConfirmTripModal} className="bg-lime-300 text-lime-950 rounded-lg px-5 py-2 font-medium flex items-center gap-2 hover:bg-lime-400" >
+                                Confirmar viagem <ArrowRight className="size-5" />
+                            </button>
+                        </div>
+
+                    )}
+                </div>
+                <p className="text-sm text-zinc-500 ">Ao planejar sua viagem pela plann.er você automaticamente concorda <br />
+                    com nossos <a className="text-zinc-300 underline" href="#">termos de uso</a> e <a className="text-zinc-300 underline" href="#">políticas de privacidade</a>.</p>
+            </div>
+
+            {isGuestsModalOpen && (
+                <InviteGuestsModal 
+                addNewEmailToInvite={addNewEmailToInvite} 
+                closeGuestsModal={closeGuestModal}
+                emailsToInvite={emailsToInvite}
+                removeEmailFromInvite={removeEmailFromInvite}
+                />
+            )}
+
+
+            {isConfirmTripModalOpen && (
+               <ConfirmTripModal 
+               createTrip={createTrip}
+               closeConfirmTripModal={closeConfirmTripModal}/>
+            )}
+
+
+        </div>
+    )
+}
+
